+++ Packages/Core/EditorAssetUtility/Editor/BFMenuPriority.cs
@@ -11,25 +11,7 @@ namespace BFTools.Core.EditorAssetUtility.Editor
         }
         public static class Module
         {
-            // Core
             public const int Logger = 4;
-
-            // Systems
-            public const int GlobalBootstrapper = 1;
-            public const int LevelBootstrapper = 2;
-            public const int ObjectPooler = 3;
-            public const int SceneManager = 5;
-
-            // Feedback
-            public const int Haptics = 1;
-            public const int ScreenShake = 2;
-            public const int Hitstop = 3;
-            public const int ScreenFlash = 4;
-
-            // Visuals
-            public const int Palette = 1;
-            public const int Background = 2;
-            public const int Parallax = 3;
         }
     }
 }