+++ Packages/Feedback/ScreenFlash/Editor/BFScreenFlashConfigCreator.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using BFTools.Core.EditorAssetUtility.Editor;
+using BFTools.Feedback.EditorAssetUtility.Editor;
 
 namespace BFTools.Feedback.ScreenFlash.Editor
 {
@@ -8,7 +9,7 @@ namespace BFTools.Feedback.ScreenFlash.Editor
         private const string TargetPath = "Assets/Configs/Feedback/ScreenFlash";
         private const string AssetName = "ScreenFlashConfig.asset";
 
-        [MenuItem("Assets/Create/BFTools/Feedback/Config/Screen Flash Config", priority = BFMenuPriority.Group.Feedback + BFMenuPriority.Module.ScreenFlash)]
+        [MenuItem("Assets/Create/BFTools/Feedback/Config/Screen Flash Config", priority = BFFeedbackMenuPriority.ScreenFlash)]
         private static void Create()
         {
             BFEditorAssetUtility.CreateConfigAsset<BFScreenFlashConfig>(TargetPath, AssetName);