+++ Packages/Feedback/ScreenShake/Editor/BFScreenShakeConfigCreator.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using BFTools.Core.EditorAssetUtility.Editor;
+using BFTools.Feedback.EditorAssetUtility.Editor;
 
 namespace BFTools.Feedback.ScreenShake.Editor
 {
@@ -8,7 +9,7 @@ namespace BFTools.Feedback.ScreenShake.Editor
         private const string TargetPath = "Assets/Configs/Feedback/ScreenShake";
         private const string AssetName = "ScreenShakeConfig.asset";
 
-        [MenuItem("Assets/Create/BFTools/Feedback/Config/Screen Shake Config", priority = BFMenuPriority.Group.Feedback + BFMenuPriority.Module.ScreenShake)]
+        [MenuItem("Assets/Create/BFTools/Feedback/Config/Screen Shake Config", priority = BFFeedbackMenuPriority.ScreenShake)]
         private static void Create()
         {
             BFEditorAssetUtility.CreateConfigAsset<BFScreenShakeConfig>(TargetPath, AssetName);