+++ Packages/Systems/GlobalBootstrapper/Editor/BFGlobalBootstrapConfigCreator.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using BFTools.Core.EditorAssetUtility.Editor;
+using BFTools.Systems.EditorAssetUtility.Editor;
 
 namespace BFTools.Systems.GlobalBootstrapper.Editor
 {
@@ -8,7 +9,7 @@ namespace BFTools.Systems.GlobalBootstrapper.Editor
         private const string TargetPath = "Assets/Resources/BFTools";
         private const string AssetName = "GlobalBootstrapConfig.asset";
 
-        [MenuItem("Assets/Create/BFTools/Systems/Config/Global Bootstrap Config", priority = BFMenuPriority.Group.Systems + BFMenuPriority.Module.GlobalBootstrapper)]
+        [MenuItem("Assets/Create/BFTools/Systems/Config/Global Bootstrap Config", priority = BFSystemsMenuPriority.GlobalBootstrapper)]
         private static void Create()
         {
             BFEditorAssetUtility.CreateConfigAsset<BFGlobalBootstrapperConfig>(TargetPath, AssetName);