+++ Packages/Systems/ObjectPooler/Editor/BFObjectPoolerVariantCreator.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using BFTools.Core.EditorAssetUtility.Editor;
+using BFTools.Systems.EditorAssetUtility.Editor;
 
 namespace BFTools.Systems.ObjectPooler.Editor
 {
@@ -9,7 +10,7 @@ namespace BFTools.Systems.ObjectPooler.Editor
         private const string TargetPath = "Assets/Prefabs/Systems";
         private const string AssetName = "ObjectPooler.prefab";
 
-        [MenuItem("Assets/Create/BFTools/Systems/Prefabs/Object Pooler", priority = BFMenuPriority.Group.Systems + BFMenuPriority.Module.ObjectPooler)]
+        [MenuItem("Assets/Create/BFTools/Systems/Prefabs/Object Pooler", priority = BFSystemsMenuPriority.ObjectPooler)]
         private static void Create()
         {
             BFEditorAssetUtility.CreatePrefabVariant(BasePrefabPath, TargetPath, AssetName);