+++ Packages/Systems/SceneManager/Editor/BFSceneLoadRequestCreator.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using BFTools.Core.EditorAssetUtility.Editor;
+using BFTools.Systems.EditorAssetUtility.Editor;
 
 namespace BFTools.Systems.SceneManager.Editor
 {
@@ -8,7 +9,7 @@ namespace BFTools.Systems.SceneManager.Editor
         private const string TargetPath = "Assets/Configs/Systems/SceneManager";
         private const string AssetName = "SceneLoadRequest.asset";
 
-        [MenuItem("Assets/Create/BFTools/Systems/Config/Scene Load Request", priority = BFMenuPriority.Group.Systems + BFMenuPriority.Module.SceneManager)]
+        [MenuItem("Assets/Create/BFTools/Systems/Config/Scene Load Request", priority = BFSystemsMenuPriority.SceneManager)]
         private static void Create()
         {
             BFEditorAssetUtility.CreateConfigAsset<BFSceneLoadRequest>(TargetPath, AssetName);