+++ Packages/Visuals/Background/Editor/BFTwinklingStarsLayerConfigCreator.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using BFTools.Core.EditorAssetUtility.Editor;
+using BFTools.Visuals.EditorAssetUtility.Editor;
 
 namespace BFTools.Visuals.Background.Editor
 {
@@ -8,7 +9,7 @@ namespace BFTools.Visuals.Background.Editor
         private const string TargetPath = "Assets/Configs/Visuals/Background/Layers/TwinklingStars";
         private const string AssetName = "TwinklingStarsLayerConfig.asset";
 
-        [MenuItem("Assets/Create/BFTools/Visuals/Background/Config/Layer/Twinkling Stars Layer Config", priority = BFMenuPriority.Group.Visuals + BFMenuPriority.Module.Background)]
+        [MenuItem("Assets/Create/BFTools/Visuals/Background/Config/Layer/Twinkling Stars Layer Config", priority = BFVisualsMenuPriority.Background)]
         private static void Create()
         {
             BFEditorAssetUtility.CreateConfigAsset<BFTwinklingStarsLayerConfig>(TargetPath, AssetName);