+++ Packages/Visuals/Parallax/Editor/BFParallaxSpriteLayerConfigCreator.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using BFTools.Core.EditorAssetUtility.Editor;
+using BFTools.Visuals.EditorAssetUtility.Editor;
 
 namespace BFTools.Visuals.Parallax.Editor
 {
@@ -8,7 +9,7 @@ namespace BFTools.Visuals.Parallax.Editor
         private const string TargetPath = "Assets/Configs/Visuals/Parallax/Layers/Sprite";
         private const string AssetName = "ParallaxSpriteLayerConfig.asset";
 
-        [MenuItem("Assets/Create/BFTools/Visuals/Parallax/Config/Layer/Sprite Layer Config", priority = BFMenuPriority.Group.Visuals + BFMenuPriority.Module.Parallax)]
+        [MenuItem("Assets/Create/BFTools/Visuals/Parallax/Config/Layer/Sprite Layer Config", priority = BFVisualsMenuPriority.Parallax)]
         private static void Create()
         {
             BFEditorAssetUtility.CreateConfigAsset<BFParallaxSpriteLayerConfig>(TargetPath, AssetName);